Move folder user from entities to shared and update imports
