+++ src/entities/post/ui/postItem/PostItem.tsx
@@ -1,6 +1,6 @@
 import { Post } from '@/features/post/types/postTypes'
 import { DateStamp } from '@/shared/ui'
-import { UserCard } from '@/entities/user'
+import { UserCard } from '@/shared/user'
 
 import s from './PostItem.module.scss'
