+++ src/shared/user/ui/userCard/UserCard.module.scss
